Refactored to remove the utilities from toluene base.
