+++ toluene/__init__.py
@@ -1,2 +1,3 @@
 import toluene.base
+import toluene.elevation
 import toluene.test
